requirements_minimal.txt
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
yfinance>=0.2.0
scikit-learn>=1.0.0
pyyaml>=6.0
seaborn>=0.11.0
gymnasium>=0.29.1
